requirements-lite.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tensorflow-cpu>=2.8.0
scikit-learn>=1.0.2
pandas>=1.4.0
numpy>=1.21.0
fastapi-limiter>=0.1.5
python-dotenv>=0.19.0
pydantic>=1.9.0
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.28.0
python-multipart>=0.0.5
tenacity>=8.0.1 